refactor: adjust path to auth module so that it always works from working directory
